PLTR as default stock example
The default stock used for analysis is Palantir. it is worth noting that the algorithm does not take into account sentiment analysis and as such it is wise to pick a time frame for the data that does not encapsulate dramatic real world events such as Covid-19 or the housing crisis in 2009. With original testing, the 10 year period between the 2 crises was used for testing however, the model proved most effective over a 1 year time frame.